requirements.txt
apscheduler==3.7.0
discord.py==1.7.3
hashids==1.3.1 
matplotlib==3.4.2
pytz==2021.1